Fix rebalance background job ID type

+++ src/backend/distributed/operations/shard_rebalancer.c
@@ -1201,7 +1201,7 @@ citus_rebalance_start(PG_FUNCTION_ARGS)
 		.rebalanceStrategy = strategy,
 		.improvementThreshold = strategy->improvementThreshold,
 	};
-	int jobId = RebalanceTableShardsBackground(&options, shardTransferModeOid,
+	int64 jobId = RebalanceTableShardsBackground(&options, shardTransferModeOid,
 											   ParallelTransferReferenceTables,
 											   ParallelTransferColocatedShards);
